GPU TPC: Use unsigned int for occupancy bin debug streamer

+++ GPU/GPUTracking/Merger/GPUTPCGMMergerDump.cxx
@@ -270,9 +270,9 @@ void GPUTPCGMMerger::DebugRefitMergedTrack(const GPUTPCGMMergedTrack& track) con
   }
 }
 
-std::vector<unsigned short> GPUTPCGMMerger::StreamerOccupancyBin(int iSlice, int iRow, float time) const
+std::vector<unsigned int> GPUTPCGMMerger::StreamerOccupancyBin(int iSlice, int iRow, float time) const
 {
-  std::vector<unsigned short> retVal(1 + 2 * Param().rec.tpc.occupancyMapTimeBinsAverage);
+  std::vector<unsigned int> retVal(1 + 2 * Param().rec.tpc.occupancyMapTimeBinsAverage);
 #ifdef DEBUG_STREAMER
   const int bin = CAMath::Max(0.f, time / Param().rec.tpc.occupancyMapTimeBins);
   for (int i = 0; i < 1 + 2 * Param().rec.tpc.occupancyMapTimeBinsAverage; i++) {
